models/requirements.txt
flask
langchain
langchain-community
google-generativeai
python-dotenv
ipykernel
streamlit
requests
PyPDF2
python-decouple
pydub
torch
transformers